+++ src/main/java/com/ormi5/movieblog/postcontroller/PostRepository.java
@@ -1,9 +1,9 @@
-package com.ormi5.movieblog.repository;
+package com.ormi5.movieblog.postcontroller;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-import com.ormi5.movieblog.entity.Post;
+import com.ormi5.movieblog.post.Post;
 
 @Repository
 public interface PostRepository extends JpaRepository<Post, Long> {
